requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
requests==2.31.0
